+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.36.0] - 2025-08-16
+## [1.36.0] - 2025-08-17
 
 ### Changed
 - Support `SYNC_WRITE` event in the event callback system.  
   This event is triggered when an object (or part of an object) is written to a target.
+- Support `SYNC_FILTERED` event in the event callback system.  
+  If an object is filtered out, this event is triggered. EventData.message will contain the reason for filtering.
 - Updated dependencies.
 
 ## [1.35.0] - 2025-08-15